Add libm dependency for no_std
This commit adds the `libm` crate as an optional dependency, primarily
to support `no_std` environments. It enables the `no_std` feature in
`Cargo.toml` to conditionally include `libm`. An example
`alloc_nostd.rs` is also included to demonstrate `kramaframe` usage in a
no_std environment with the `alloc` feature.